+++ NEWS.rst
@@ -6,6 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
+* Major: Renamed ``read`` to ``read_bytes``.
+* Major: Renamed ``peek`` to ``peek_bytes``.
 * Major: Upgraded to endian 8.
 * Major: Upgraded to bitter 4.